feat: [KanuKim97]: 점주 상품 등록 및 찜 관리 API 연동 (#99)
* feat: [KanuKim97]: 점포 관리 구현

* feat: [KanuKim97]: 상품 관리 수정

* feat: [KanuKim97]: 점주 상품 및 찜 관리 API 연동

* chore: [KanuKim97]: 9/18 OpenAPI 상품 목록 최신화

* feat: [KanuKim97]: 태그 제거 및 사진 첨부 1장으로 구현 및 API 수정

* fix: [KanuKim97]: 상품 관리 디자인 테스트 Paging fixture 수정

+++ feature/owner/product/src/main/kotlin/com/swyp/mangro/feature/owner/product/screen/editor/ProductRegistrationViewModel.kt
@@ -93,8 +93,8 @@ class ProductRegistrationViewModel internal constructor(
                             salePrice = saved.salePrice,
                             initialQuantity = saved.initialQuantity,
                             remainingQuantity = saved.stockQuantity,
-                            reservedQuantity = saved.heldQuantity,
-                            pickedUpQuantity = saved.completedQuantity,
+                            reservedQuantity = saved.heldQuantity.toLong(),
+                            pickedUpQuantity = saved.completedQuantity.toLong(),
                             pickupEndTime = OffsetDateTime.parse(saved.pickupEndAt).atZoneSameInstant(zone).toLocalTime().toString().take(5),
                             tags = emptyList(),
                         )
